Changes to flow diagrams
Lots of small visualisation changes to the FlowDiagram generator. Mostly
CSS, but the flowdiagram xsl now outputs StatementItems as well - these
show up as purple in the tool. Some unneccessary spans that were
crufting up the CSS are now gone to.
The DogSurvey now has a Statement to make it a more rounded test case.