Changes for instructions and subquestion numbering
Subquestion Numbering is now done as a part of the CSS using ordered
lists, the theme config XSD and config files and have changed to
accomodate this. Because of block elements wrecking the display of
lists, the question for the main numbers is still done in the XSLT.

A few slight changes to the Dog Survey to fix some validation errors.

The Ramona Theme will need to be changed accordingly as its probably now
horribly ugly(ier).